refactor(net-stubbing): replace driver adapters with CyIntercept class
