--- packages/net-stubbing/lib/adapters/index.ts
@@ -1,5 +1,0 @@
-export { DriverInterceptionEventsAdapter } from './driver-interception-events-adapter'
-
-export { createDriverAdapter } from './create-driver-adapter'
-
-export type { DriverAdapter } from './create-driver-adapter'
