+++ packages/net-stubbing/lib/index.ts
@@ -1,0 +1,3 @@
+export { CyIntercept, INTERCEPT_HEADERS, _restoreMatcherOptionsTypes } from './cy-intercept'
+
+export type { CyInterceptConfig, CyInterceptOptions, DriverEventFrame } from './cy-intercept'
